Refactor script for better usability
This commit improves the usability of this utility. Now the flags are
preprocessed to check for the help (-h/--help etc.) flag; before this
was poorly handled as the password was required and the help message was
printed once for every camera. Also it is determined if the alarm is
being armed or disarmed. If the system is being armed there is a 30
second grace period for the user to exit the premises.